Bad token of activity when do the launch work at the app process side.
If the activity destroy timeout, AMS will clear it from the task.

And if activity LAUNCH_ACTIVITY message is blocked in the app main thread,
it may do the handleLaunchActivity after AMS removed the activity record
token from the task.At this time, ActivityThread launch the activity and
call some AMS interfaces like reportSizeConfigurations(isTopOfTask etc.),
AMS may throw IllegalArgumentException for the bad token.

Good for the system stability, we should check if the launching activity
is in any stack to deceide whether need to do the follow-up launch work.