1 SystemUI: disable statusbar time refresh when screen off
When screen is off, don't refresh the status bar to save power.
When screen is on, refresh immediately for user experience.
 2 SystemUI: force a clock update if the screen is turned back on
The statusbar time must be updated immediately when the screen turns back on
without depending on ACTION_TIME intent to be called later.
 3 SystemUI: move time tick to bg handler
When system is heavy, the updateClock time tick will result in
ANR. Move it to bg handler to avoid it.