Fix shared libraries loading during backup
When backup or restore is triggered on a non running
application the shared libraries are not loaded. This
will result in ClassNotFoundException when launching
an application referenced a shared library. Fix is to
add a GET_SHARED_LIBRARIES flag when calling
getApplicationInfo, then loading of shared library
when application launch is triggered by backup.