Native crash occurs. android.renderscript.FieldPacker#subalign()
Bug Details:
Although arguments are checked by android.renderscript.FieldPacker#subalign(),
Native crash occurs without RSIllegalArgumentException being thrown because there is no consideration when the argument is 0 or less.

Modification:
When checking the argument of the subalign() method,
if it is 0 or less (like align()), a condition to throw RSIllegalArgumentException is added.