Do not force crash for EGL_NO_SURFACE
A forced crash will happen if an UI object is removed right after it
is created and the OpenGLRenderer tries to render the removed object.
A system crash and a reboot if this occurs in the system_server.
Removing the forced crash and instead return the error code
EGL_NO_SURFACE is a better option. The responsibilty to handle this is
moved to the caller.