KernelCpuSpeedReader: Account for missing sysfs nodes
It is possible that after CPU/cluster hotplugging
the cpufreq stats sysfs nodes don't get recreated.

In that case we would be encountering a lot of failed
read attempts which are obsolete.

As solution, this adds a detection mechanism
which checks the cpu online status and implements
a failure counter.
The failure counter gets incremented every time
a read attempt to the stats sysfs node fails while
the cpu was online. This case is considered as proof for
the sysfs bug.
Once the failure counter reaches MAX_READ_TRIES (3 by default),
no more read attempts are made.

Change-Id: Id233077797da65aa09e63a964e7cf4d96179c92f