Wifi: Check for WiFiService's existence before its access
SystemUI , currently accesses Wi-FiService's API's before it is initialized/
started.The current behaviour is to not start the Wi-Fi Service during the
FDE encryption ( Wi-FI Service is not considered as a coreService). This
behavior would lead to a fatal exception. Address this by adding a NULL check
before the access in all the components of systemUI.

Change-Id: I5c375ab99760847987a3fd5602cf1195725c74fa