Return null if a provider died before acquisition
In case of a race between appDiedLocked and getContentProvider, it was
possible for a holder with a null provider to be returned.  The caller
of getContentProvider would then typically try to instantiate the
provider locally, and fail with a security exception.

Instead of returning an invalid holder with a null provider, return a
null holder.  getContentProvider already returns null in several cases,
so callers should be prepared for it.

Change-Id: If89c3260ddb65221c10285b8301b984b583b0c7d