Make services with uid 1000 restart if their host process crashes
The framework will only start a pendding service only if the service
has the same uid with the host process, this is usually right. But
there's an exception, the system uid (uid 1000) can make their
components run on another process without having the same uid. Services
like this will stay in the pendding services list but be ignored
when they need to start.

Change-Id: I5fc80489dc7590845395121b171a87632dcae1fe