Clean receivers completely when forceStop the process
when forceStop is called, then the broadcast-receiver will be
cleaned up.However, only receivers of the userId is cleaned.
When the userId is not UserHandle.USER_ALL and the userId of
BroadcastRecord is UserHandle.USER_ALL. Then the receiver of
this BR will not be cleaned up. Thus the receiver of this
process with this userId will be dispatched. What's worse, it
may cause severe problems under multi-user

Change-Id: I2b26db9c0ce4517245b81396563e8daac05cc684