OMS: Dismantle FLAG_OVERLAY_TRUSTED check in isOverlayPackage
This was added with commit 7a722dcae9ac but this appears to be an
incomplete attempt at doing what it should, because FLAG_OVERLAY_TRUSTED
is never assigned, only checked. While we wait for Google to realise this
and push a fuller commit for us to mitigate against, this bandaid does the
trick.

Change-Id: Ibbfebb7617515e50c909f45ab23d87724cfc4197