Expose and tint settings dashboard icon tint
Doing some research, I found that settings icons are tinted by using
colorControlNormal. While using our dark theme, we define colorControlNormal
and call on the system accent. Everything looks great using the dark theme.

The problem here was that the system was not defining it in the styles while
using the light(default) theme so the icon tint default to
737373 so when
you applied an accent using the light theme, those icons were not being changed.

Change-Id: I0dd195f9c07ac45380d3bc9b8c9d902f1a43c4bd